+++ attributes/default.rb
@@ -1,0 +1,6 @@
+default["OGR-28"] = true
+default["RC-72"] = true
+default["dnetc_url"] = "http://http.distributed.net/pub/dcti/current-client/dnetc-linux-arm-eabi.tar.gz"
+default["dnetc_pgk"] = "dnetc-linux-arm-eabi.tar.gz"
+default["dnetc_cache_dir"] = "/tmp"
+default["dnetc_install_dir"] = "/usr/local/dnetc"