+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This file is used to list changes made in each version of the pi-dnetc cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of pi-dnetc
+- [Andreas Fritzler] - Initial release of pi-dnetc
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.